requirements.txt
pytesseract
Pillow
SpeechRecognition
pydub
pyogg